refactor: 공용 MemberAvatar 컴포넌트 분리
